fix bug where articles in a blog would disappear if the same was in a resteem feed and vice versa. Hide images by default. android studio lint optimizations

+++ app/src/main/java/com/steemapp/lokisveil/steemapp/RoomDatabaseApp/RoomDaos/ArticleDao.kt
@@ -29,11 +29,11 @@ interface ArticleDao {
 
     //get an item matching article id
     @Query("SELECT * from article_holder where id == :eid")
-    fun getArticle(eid:Int): LiveData<FeedArticleDataHolder.FeedArticleHolder>
+    fun getArticle(eid:Long): LiveData<FeedArticleDataHolder.FeedArticleHolder>
 
     //get an item matching article id
     @Query("SELECT * from article_holder where id == :eid")
-    fun getArticleNormal(eid:Int): FeedArticleDataHolder.FeedArticleHolder?
+    fun getArticleNormal(eid:Long): FeedArticleDataHolder.FeedArticleHolder?
 
     //get an article matching internal db
     @Query("SELECT * from article_holder where myDbKey == :eid")
@@ -41,15 +41,15 @@ interface ArticleDao {
 
     //get a paged list of all items
     @Query("SELECT id,myDbKey,active,activeVotes,already_paid,app,author,authorreputation,cashoutTime,category,children,created,createdcon,date,datespan,depth,displayName,entryId,followsYou,format,image,lastPayout,lastUpdate,links,pending_payout_value,netVotes,permlink,promoted,reblogBy,reblogOn,replies,rootAuthor,rootComment,rootPermlink,summary,tags,title,total_pending_payout_value,useFollow,users,uservoted,width,isBlog,saveTime from article_holder where isBlog == :isBlog ORDER BY myDbKey ASC")
-    fun getPagedList(isBlog:Boolean = false): DataSource.Factory<Integer,FeedArticleDataHolder.FeedArticleHolder>
+    fun getPagedList(isBlog:Boolean = false): DataSource.Factory<Int,FeedArticleDataHolder.FeedArticleHolder>
 
     //get a paged list of all items ordered by savetime
     @Query("SELECT id,myDbKey,active,activeVotes,already_paid,app,author,authorreputation,cashoutTime,category,children,created,createdcon,date,datespan,depth,displayName,entryId,followsYou,format,image,lastPayout,lastUpdate,links,pending_payout_value,netVotes,permlink,promoted,reblogBy,reblogOn,replies,rootAuthor,rootComment,rootPermlink,summary,tags,title,total_pending_payout_value,useFollow,users,uservoted,width,isBlog,saveTime from article_holder where isBlog == :isBlog ORDER BY saveTime DESC")
-    fun getPagedListTime(isBlog:Boolean = false): DataSource.Factory<Integer,FeedArticleDataHolder.FeedArticleHolder>
+    fun getPagedListTime(isBlog:Boolean = false): DataSource.Factory<Int,FeedArticleDataHolder.FeedArticleHolder>
 
     //get a paged list after an id
     @Query("SELECT id,myDbKey,active,activeVotes,already_paid,app,author,authorreputation,cashoutTime,category,children,created,createdcon,date,datespan,depth,displayName,entryId,followsYou,format,image,lastPayout,lastUpdate,links,pending_payout_value,netVotes,permlink,promoted,reblogBy,reblogOn,replies,rootAuthor,rootComment,rootPermlink,summary,tags,title,total_pending_payout_value,useFollow,users,uservoted,width,isBlog,saveTime from article_holder where isBlog == :isBlog and myDbKey > :dbkey ORDER BY myDbKey ASC")
-    fun getPagedList(dbkey:Int,isBlog:Boolean = false): DataSource.Factory<Integer,FeedArticleDataHolder.FeedArticleHolder>
+    fun getPagedList(dbkey:Int,isBlog:Boolean = false): DataSource.Factory<Int,FeedArticleDataHolder.FeedArticleHolder>
 
     //get the last items db id
     @Query("SELECT myDbKey FROM article_holder WHERE myDbKey = (SELECT MAX(myDbKey) FROM article_holder)")
